Fix definition of pilling job's parameters.
In this plugin default values of polling job's parameters
were not overriden but new parameters of String type
were created in any cases.

Mechanism that replaces current one was borrowed from plugin
https://github.com/jenkinsci/parameterized-scheduler-plugin

We updated Jenkins Plugin version and after that we had to add
dependencies in this project that had been moved from Jenkins
Core to separate plugins.

task:GENCM2022-3