Add ability to set/override values of job parameters
Could be used in Scheduler field as a commented line.
In this field comments follow a sharp sign (#).
So start each param line with the sign.

Syntax:
  @param <name> [value]

Parameters are applied to a cron item above which they are listed.
These parameters values override ones of default job parameters.
No quotes are needed in case a parameter value contains whitespaces.
A parameter value may be omitted. In this case it will be assumed as an empty one.